- exiftool gem added for checking the width and height of video to get orientation - mini_exiftool added

+++ app/jobs/convert_portrait_to_landscape_job.rb
@@ -5,14 +5,18 @@ class ConvertPortraitToLandscapeJob < ApplicationJob
     if order.present? && video_count > 0
       video_count.times do |index|
         video_path = ActiveStorage::Blob.service.path_for(order&.videos[index]&.key)
-        temp_file = "#{Rails.root}/testing#{index}.mp4"
-        system(`ffmpeg -i "#{video_path}" -lavfi "color=color=black@.5:size=270x360:d=1[dark];[0:v]crop=270:360[blurbase];    [blurbase]boxblur=luma_radius='min(h,w)/20':luma_power=1:chroma_radius='min(cw,ch)/20':chroma_power=1[blurred];  [blurred][dark]overlay[darkened]; [darkened]scale=640:360[bg];    [0:v]scale=-1:360[fg]; [bg][fg]overlay=(W-w)/2:(H-h)/2" "testing#{index}.mp4"`)
-        if (File.exists?(temp_file))
-          downloaded_video = open(temp_file)
-          order.videos.attach(io: downloaded_video  , filename: "#{order&.videos[index]&.filename}")
-          order.videos[index].purge
-          downloaded_video.close
-          File.delete(temp_file)  
+        video_data = MiniExiftool.new("#{video_path}")
+        if video_data.imageheight > video_data.imagewidth
+          temp_file = "#{Rails.root}/testing#{index}.mp4"
+          portrait_cropped = "#{Rails.root}/portrait_cropped.mp4"
+          system(`ffmpeg -i "#{video_path}" -lavfi "color=color=black@.5:size=270x360:d=1[dark];[0:v]crop=270:360[blurbase];    [blurbase]boxblur=luma_radius='min(h,w)/20':luma_power=1:chroma_radius='min(cw,ch)/20':chroma_power=1[blurred];  [blurred][dark]overlay[darkened]; [darkened]scale=640:360[bg];    [0:v]scale=-1:360[fg]; [bg][fg]overlay=(W-w)/2:(H-h)/2" "testing#{index}.mp4"`)
+          if (File.exists?(temp_file))
+            downloaded_video = open(temp_file)
+            order.videos.attach(io: downloaded_video  , filename: "#{order&.videos[index]&.filename}")
+            order.videos[index].purge
+            downloaded_video.close
+            File.delete(temp_file)  
+          end
         end
       end
     end
